cardio-back/src/main/java/com/sopra/agile/cardio/back/model/DbSprint.java
package com.sopra.agile.cardio.back.model;

import org.joda.time.LocalDate;

public class DbSprint {

    private long id;
    private String name;
    private LocalDate startDate;
    private LocalDate endDate;
    private String goal;
    private int commitment;
    private int velocity;

    public DbSprint() {
        // Empty constructor
    }

    public DbSprint(long id, final String name, final LocalDate startDate, final LocalDate endDate) {
        this.id = id;
        this.name = name;
        this.startDate = startDate;
        this.endDate = endDate;
    }

    public long getId() {
        return id;
    }

    public void setId(long id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public LocalDate getStartDate() {
        return startDate;
    }

    public void setStartDate(LocalDate startDate) {
        this.startDate = startDate;
    }

    public LocalDate getEndDate() {
        return endDate;
    }

    public void setEndDate(LocalDate endDate) {
        this.endDate = endDate;
    }

    public String getGoal() {
        return goal;
    }

    public void setGoal(String goal) {
        this.goal = goal;
    }

    public int getCommitment() {
        return commitment;
    }

    public void setCommitment(int commitment) {
        this.commitment = commitment;
    }

    public int getVelocity() {
        return velocity;
    }

    public void setVelocity(int velocity) {
        this.velocity = velocity;
    }
}
